pyproject.toml
[project]
name = "ai-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.11"
dependencies = [
    "fastapi>=0.129.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.13",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.10",
    "pydantic-settings>=2.13.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
    "langchain-pinecone>=0.2.0",
]
